+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
-## Unreleased 
+## Unreleased
 ### Add
+   - Unify filename creation for feed export file
    - Category Page & Search Result Page
       - add Scroll to top callback after products page change